Moving files around so they are discoverable by gradle tests
