+++ src/main/java/xbot/common/subsystems/drive/MockDriveSubsystem.java
@@ -6,7 +6,6 @@ import java.util.Map;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
 
-import xbot.common.subsystems.BaseDriveSubsystem;
 import xbot.common.controls.actuators.XCANTalon;
 import xbot.common.injection.wpi_factories.CommonLibFactory;
